Change the type of phoneNum, fix type errors

+++ backend/src/main/java/com/laioffer/flagcamp/backend/controller/PostController.java
@@ -22,7 +22,7 @@ public class PostController {
 
     // GET /api/posts -> 获取所有帖子 (支持按标签筛选)
     @GetMapping
-    public ResponseEntity<List<Post>> getAllPosts(@RequestParam(required = false) Integer tagId) {
+    public ResponseEntity<List<Post>> getAllPosts(@RequestParam(required = false) Long tagId) {
         List<Post> posts = postService.getAllPosts(Optional.ofNullable(tagId));
         return ResponseEntity.ok(posts);
     }
@@ -36,7 +36,7 @@ public class PostController {
 
     // POST /api/posts -> 发布帖子 (绑定商品 + 标签)
     @PostMapping
-    public ResponseEntity<Post> createPost(@RequestBody CreatePostRequest request, @RequestParam Integer ownerId) {
+    public ResponseEntity<Post> createPost(@RequestBody CreatePostRequest request, @RequestParam Long ownerId) {
         // ‼️ 安全警告: 在生产环境中，ownerId 必须从安全认证信息(如JWT Token)中获取，
         // 而不是通过 @RequestParam 从 URL 参数传递，否则任何用户都可以冒充他人发帖。
         Post newPost = new Post(null, request.postedItemId(), request.tagId(), ownerId);
@@ -46,7 +46,7 @@ public class PostController {
 
     // DELETE /api/posts/{id} -> 删除帖子
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long id, @RequestParam Integer userId) {
+    public ResponseEntity<Void> deletePost(@PathVariable Long id, @RequestParam Long userId) {
         // ‼️ 安全警告: 同上，userId 必须从安全认证信息中获取，以验证操作权限。
         postService.deletePost(id, userId);
         return ResponseEntity.noContent().build();
